+++ src/main/java/io/fabric8/maven/docker/AbstractDockerMojo.java
@@ -119,10 +119,6 @@ public abstract class AbstractDockerMojo extends AbstractMojo implements Context
     @Parameter(property = "docker.skip", defaultValue = "false")
     private boolean skip;
 
-    // Max number of containers to fetch when stopping/removing
-    @Parameter(property = "docker.fetchLimit", defaultValue = "100")
-    protected int fetchLimit;
-
     /**
      * Whether the usage of docker machine should be skipped competely
      */
@@ -276,7 +272,6 @@ public abstract class AbstractDockerMojo extends AbstractMojo implements Context
                 access = new DockerAccessWithHcClient("v" + version, dockerUrl,
                                                       dockerConnectionDetector.getCertPath(certPath),
                                                       maxConnections,
-                                                      fetchLimit,
                                                       log);
                 access.start();
                 setDockerHostAddressProperty(dockerUrl);